requirements-dev.txt
codecov==2.0.15
flake8==3.7.0
mypy==0.641
nbval==0.9.1
numpydoc==0.8.0
pylint==2.2.3
pytest==3.9.1
pytest-cov==2.6.0
sphinx==2.0
sphinx-gallery==0.3.1
twine==1.14.0
yapf==0.26.0
